--- semestralni_prace/coworking/aboutus.php
@@ -1,35 +1,0 @@
-<?php include __DIR__ . '/includes/header.php'; ?>
-
-
-<div class="bg-light">
-    <div class="container py-5">
-        <div class="row h-100 align-items-center py-5">
-            <div class="col-lg-6">
-                <h3 class="display-5">About us page</h3>
-                <p class="lead text-muted mb-0">Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
-            </div>
-            <div class="col-lg-6 d-none d-lg-block"><img src="https://res.cloudinary.com/mhmd/image/upload/v1556834136/illus_kftyh4.png" alt="" class="img-fluid"></div>
-        </div>
-    </div>
-</div>
-
-<div class="bg-white py-5">
-  <div class="container py-5">
-    <div class="row align-items-center mb-5">
-      <div class="col-lg-6 order-2 order-lg-1"><i class="fa fa-bar-chart fa-2x mb-3 text-primary"></i>
-        <h3 class="font-weight-light">Lorem ipsum dolor sit amet</h3>
-        <p class="font-italic text-muted mb-4">Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p><a href="#" class="btn btn-light px-5  shadow-sm">Learn More</a>
-      </div>
-    </div>
-    <div class="row align-items-center">
-      <div class="col-lg-5 px-5 mx-auto"></div>
-      <div class="col-lg-6">
-        <h3 class="font-weight-light">Lorem ipsum dolor sit amet</h3>
-        <p class="font-italic text-muted mb-4">Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p><a href="#" class="btn btn-light px-5  shadow-sm">Learn More</a>
-      </div>
-    </div>
-  </div>
-</div>
-
-
-<?php include __DIR__ . '/includes/footer.php'; ?>
