--- semestralni_prace/coworking/db.php
@@ -1,9 +1,0 @@
-<?php
-$pdo = new PDO(
-    'mysql:host=localhost;dbname=coworking;charset=utf8mb4', 
-    'root', 
-    ''
-);
-$pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-$pdo->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
-?>
