--- semestralni_prace/coworking/pricing.php
@@ -1,57 +1,0 @@
-<?php include __DIR__ . '/includes/header.php'; ?>
-
-<br></br>
-
-<section class="pricing py-5">
-    <div class="container">
-    <h3>Pricing</h3>
-
-        <div class="row">
-            <!-- Free Tier -->
-            <div class="col-lg-6">
-                <div class="card mb-5 mb-lg-0">
-                    <div class="card-body">
-                        <h5 class="card-title text-muted text-uppercase text-center">Free</h5>
-                        <h6 class="card-price text-center">$0<span class="period">/month</span></h6>
-                        <hr>
-                        <ul class="fa-ul">
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>Single User</li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>5GB Storage</li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>Unlimited Public Projects</li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>Community Access</li>
-                            <li class="text-muted"><span class="fa-li"><i class="fas fa-times"></i></span>Unlimited Private Projects</li>
-                            <li class="text-muted"><span class="fa-li"><i class="fas fa-times"></i></span>Dedicated Phone Support</li>
-                            <li class="text-muted"><span class="fa-li"><i class="fas fa-times"></i></span>Free Subdomain</li>
-                            <li class="text-muted"><span class="fa-li"><i class="fas fa-times"></i></span>Monthly Status Reports</li>
-                        </ul>
-                        <a href="register.php" class="btn btn-light px-5  shadow-sm">Sign up</a>
-                    </div>
-                </div>
-            </div>
-            <!-- Plus Tier -->
-            <div class="col-lg-6">
-                <div class="card mb-5 mb-lg-0">
-                    <div class="card-body">
-                        <h5 class="card-title text-muted text-uppercase text-center">Plus</h5>
-                        <h6 class="card-price text-center">$9<span class="period">/month</span></h6>
-                        <hr>
-                        <ul class="fa-ul">
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span><strong>5 Users</strong></li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>50GB Storage</li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>Unlimited Public Projects</li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>Community Access</li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>Unlimited Private Projects</li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>Dedicated Phone Support</li>
-                            <li><span class="fa-li"><i class="fas fa-check"></i></span>Free Subdomain</li>
-                            <li class="text-muted"><span class="fa-li"><i class="fas fa-times"></i></span>Monthly Status Reports</li>
-                        </ul>
-                        <a href="register.php" class="btn btn-light px-5  shadow-sm">Sign up</a>
-                    </div>
-                </div>
-            </div>
-
-        </div>
-    </div>
-</section>
-
-<?php include __DIR__ . '/includes/footer.php'; ?>
