--- semestralni_prace/coworking/styles/styles.css
@@ -1,3 +1,0 @@
-.carousel-inner > .item > img {
-    height: 400px;
-  }
