--- semestralni_prace/coworking/tasks.txt
@@ -1,26 +1,0 @@
-o create database Coworking
-o create class structure
-o create home page
-    o navbar with Pricing, About us, Contact us, Login
-    o slideshow
-    o login and sign up buttons
-o create Registration page
-    o craete table Users
-    o registration form
-    o submitting registration logs you in and takes you to: client - My reservations, admin - Reservations (use session to store everything)
-    o after login new navbar: client - Profile, My reservations, Logout, admin - Reservations, Workplace, Logout
-o create Login page
-    o login form
-    o submitting login logs you in and takes you to: client - My reservations, admin - Reservations (use session to store everything)
-    o after login new navbar: client - Profile, My reservations, Logout, admin - Reservations, Workplace, Logout
-o create My reservations page
-    o create Add reservation button -> takes you to New reservation page
-    o craete Active reservations table with edit and delete buttons 
-    o create Previous reservations table 
-o create New reservation - client page
-    o craete new reservation form with dates and Submit button 
-o create Reservations page
-o create New reservation - admin page
-o create Workplaces page
-o create About us page
-o create Contact us page
