+++ www/kadm08/sp/includes/header.php
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <title>Coworking Smetana</title>
+    <link rel="shortcut icon" href="https://upload.wikimedia.org/wikipedia/commons/thumb/f/fa/Eo_circle_grey_letter-s.svg/1200px-Eo_circle_grey_letter-s.svg.png">
+    <link rel="stylesheet" href="https://bootswatch.com/4/journal/bootstrap.min.css">
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.7.1/css/all.css">
+    <link rel="stylesheet" href="/coworking/styles/styles.css">
+</head>
+
+<?php require __DIR__ . '/navbar.php'; ?>
