+++ www/kadm08/sp/includes/navbar.php
@@ -1,0 +1,39 @@
+<body>
+    <nav class="navbar navbar-expand-lg navbar-light bg-light fixed-top">
+        <div class="container">
+            <a class="navbar-brand" href="#">Coworking Smetana</a>
+            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
+                <span class="navbar-toggler-icon"></span>
+            </button>
+            <div class="collapse navbar-collapse" id="navbarResponsive">
+                <ul class="navbar-nav ml-auto">
+                    <li class="nav-item active">
+                        <a class="nav-link" href="index.php">Home
+                            <span class="sr-only">(current)</span>
+                        </a>
+                    </li>
+                     <?php if (isset($_SESSION['email'])) : ?>
+                        <li class="nav-item">
+                        <li><a class="nav-link" href="myAccount.php">My account</a>
+                        </li>
+                        <li class="nav-item">   
+                            <a class="nav-link" href="logout.php">My reservations</a>
+                        </li>
+                    <?php else : ?>
+                        <li class="nav-item">
+                            <a class="nav-link" href="pricing.php">Pricing</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link" href="aboutus.php">About us</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link" href="contact.php">Contact us</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link" href="login.php">Login</a>
+                        </li>
+                    <?php endif; ?>
+                </ul>
+            </div>
+        </div>
+    </nav>
