+++ www/kadm08/sp/styles/styles.css
@@ -1,0 +1,3 @@
+.carousel-inner > .item > img {
+    height: 400px;
+  }
